requirements.txt
# Keeper Slack App Dependencies
slack-bolt>=1.18.0
slack-sdk>=3.21.0
requests>=2.28.0
PyYAML>=6.0

